Code Refactoring and Feature Exposure
While the stl exporter supports batching, it still seemed useful to
expose STLs as an option for exporting for my workflows. A bit of code
refactoring to reduce code redundancy, and some minor clean up here and
there.

I also changed the defaults so that normals are now selected for export
by default as most folks will want to use this.

Verionsed to v0.4.5 for release