Updated for 4.2
While it's not using the new extension method for addons, etc., this
should fix issues with api calls that are no longer valid for 4.x
onward.

When I can, I'm going to migrate the addon to use the new extension
setup, but note that until that time, Blender will consider this a
legacy addon.

Signed-off-by: Steven Raybell <sraybell@gmail.com>